requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.1

# LangChain & AI/ML
langgraph==0.2.63
langchain-core>=0.3.37
langchain-google-genai==2.0.10
google-generativeai==0.8.3

# AWS & Cloud Services
boto3==1.35.95
botocore==1.35.95

# Database
snowflake-connector-python==3.12.3
psycopg2-binary==2.9.10

# PDF Generation
weasyprint==63.1

# Environment & Configuration
python-dotenv==1.0.1

# Data Models & Validation
pydantic==2.10.6

# HTTP Requests
requests==2.32.3

# CORS Middleware (included with FastAPI but explicit for clarity)
# Note: CORSMiddleware comes with FastAPI via starlette
langchain-openai>=0.1.0
openai==2.3.0
tiktoken==0.12.0

protobuf~=5.29.5
langchain~=0.3.25
idna~=3.10
httpcore~=1.0.7
httpx~=0.28.1
cryptography~=44.0.3
anyio~=4.8.0
GitPython~=3.1.45
gitdb~=4.0.12
h11~=0.14.0
defusedxml~=0.7.1
ipython~=9.3.0
numpy~=1.26.4
pillow~=11.1.0
pytz~=2025.1
toml~=0.10.2
rich~=13.9.4
ipywidgets~=8.1.5
notebook~=7.3.2
PyYAML~=6.0.2
multidict~=6.4.4
propcache~=0.3.1
typing_extensions~=4.14.1
sniffio~=1.3.1
attrs~=25.1.0
smmap~=5.0.2
click~=8.1.8
Pygments~=2.19.1
Brotli~=1.1.0
zstandard~=0.23.0
charset-normalizer~=3.4.1
setuptools~=80.9.0
overrides~=7.7.0
cffi~=1.17.1
regex~=2024.11.6
altair~=5.5.0
jsonschema~=4.23.0
narwhals~=1.48.1
pandas~=2.2.3
pyarrow~=21.0.0
Jinja2~=3.1.6
packaging~=24.2
referencing~=0.36.2
traitlets~=5.14.3
fsspec~=2025.5.1
aiohttp~=3.12.4
yarl~=1.20.0
MarkupSafe~=3.0.2
ollama~=0.4.9
orjson~=3.10.18
SQLAlchemy~=2.0.41
matplotlib~=3.10.3
scipy~=1.13.1
python-dateutil~=2.9.0.post0
pydeck~=0.9.1
xxhash~=3.5.0
aiosignal~=1.3.2
frozenlist~=1.6.0
python-multipart~=0.0.20
aiohappyeyeballs~=2.6.1
certifi~=2025.1.31
starlette~=0.41.3
pydantic_core~=2.27.2
websockets~=15.0.1
tzdata~=2025.1
pyOpenSSL~=24.3.0
tornado~=6.4.2
urllib3~=1.26.20
six~=1.17.0
greenlet~=3.2.2
psutil~=7.0.0
annotated-types~=0.7.0
tenacity~=9.1.2
langchain-community~=0.3.24
langsmith~=0.4.8
langchain-text-splitters~=0.3.8
ormsgpack~=1.10.0
anthropic~=0.59.0
requests-toolbelt~=1.0.0
cachetools~=5.5.2
blinker~=1.9.0
sympy~=1.14.0
webcolors~=24.11.1
jsonpointer~=3.0.0
uri-template~=1.3.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
jsonschema-specifications~=2024.10.1
state~=0.1.1.dev0
importlib_metadata~=8.7.0
tokenizers~=0.21.1
marshmallow~=3.26.1
safetensors~=0.5.3
torch~=2.2.2
scikit-learn~=1.6.1
filelock~=3.18.0
tqdm~=4.67.1
transformers~=4.52.3
huggingface-hub~=0.32.2
nltk~=3.9.1
beautifulsoup4~=4.13.3
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
pydantic-settings~=2.9.1
mypy_extensions~=1.1.0
google-api-python-client~=2.184.0
httpx-sse~=0.4.0
networkx~=3.4.2